Begin refactoring code
