+++ src/util/index.ts
@@ -1,0 +1,2 @@
+import { isEmptyUndOrNull, saveToFile, removeFile } from './util';
+export { isEmptyUndOrNull, saveToFile, removeFile };
